Refactor traits for vector packs
- this significantly simplifies adapting external vectorization libraries
