+++ hpx/parallel/datapar.hpp
@@ -1,0 +1,20 @@
+//  Copyright (c) 2016 Hartmut Kaiser
+//
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#if !defined(HPX_PARALLEL_DATAPAR_DATAPAR_OCT_31_2016_0308PM)
+#define HPX_PARALLEL_DATAPAR_DATAPAR_OCT_31_2016_0308PM
+
+#include <hpx/config.hpp>
+
+#if defined(HPX_HAVE_DATAPAR)
+
+#include <hpx/parallel/datapar/execution_policy.hpp>
+#include <hpx/parallel/datapar/iterator_helpers.hpp>
+#include <hpx/parallel/datapar/loop.hpp>
+#include <hpx/parallel/datapar/transform_loop.hpp>
+#include <hpx/parallel/datapar/zip_iterator.hpp>
+
+#endif
+#endif
