+++ hpx/parallel/datapar/zip_iterator.hpp
@@ -17,6 +17,7 @@
 #include <hpx/parallel/datapar/iterator_helpers.hpp>
 #include <hpx/parallel/traits/vector_pack_alignment_size.hpp>
 #include <hpx/parallel/traits/vector_pack_load_store.hpp>
+#include <hpx/parallel/traits/vector_pack_type.hpp>
 
 #include <algorithm>
 #include <iterator>
