+++ hpx/parallel/exception_list.hpp
@@ -98,19 +98,12 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 }
             }
 
-            static hpx::future<Result> call(hpx::future<Result> f)
+            static future<Result> call(future<Result> f)
             {
                 HPX_ASSERT(f.has_exception());
                 // Intel complains if this is not explicitly moved
                 return std::move(f);
             }
-
-#if defined(HPX_HAVE_DATAPAR)
-        ///////////////////////////////////////////////////////////////////////
-        template <typename Result>
-        struct handle_exception_impl<datapar_task_execution_policy, Result>
-        {
-            typedef future<Result> type;
 
             static future<Result> call(boost::exception_ptr const& outer)
             {
