+++ hpx/parallel/traits/detail/vc/vector_pack_load_store.hpp
@@ -1,4 +1,5 @@
 //  Copyright (c) 2016 Hartmut Kaiser
+//  Copyright (c) 2016 Matthias Kretz
 //
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
