+++ tests/unit/parallel/datapar_algorithms/countif_datapar.cpp
@@ -5,6 +5,7 @@
 
 #include <hpx/hpx_init.hpp>
 #include <hpx/hpx.hpp>
+#include <hpx/include/datapar.hpp>
 
 #include <iostream>
 #include <string>
